refactor: SectionsFactory로 Sections 생성 책임 위임
